backend/requirements/base.txt
django==3.2
django-cors-headers==3.7.0
djangorestframework==3.12.4
psycopg2==2.8.6
